fixes event binding mixup
The remove from order button had no type, and thus was defaulting to
type="submit".  For some reason, this was causing a keypress enter on another
field in the form to be expressed not as a submit event on the form but as a
click event on the remove button, to be handled by the click handler for that
button.  I don't know why.  Changing the button type to "button" fixes the
problems.