Don't serve "expired link" page to order owners
Basically when a user starts a pending order with a datetime within 48
hours, they can re-open the restaurant's page after the order's datetime
has been reached and instead of starting a new order, it will show the
expired link page.

With this change, we will just not select the order if there isn't an
edit_token present.